+++ R/discomb.R
@@ -527,8 +527,12 @@ discomb <- function(TE, seTE,
       stop("First argument is a pairwise object created with ",
            "'keep.all.comparisons = FALSE'.",
            call. = TRUE)
-    ##
-    seTE <- TE$seTE
+    #
+    if (is.null(attr(TE, "varnames")))
+      seTE <- TE$seTE
+    else
+      seTE <- TE[[attr(TE, "varnames")[2]]]
+    #
     treat1 <- TE$treat1
     treat2 <- TE$treat2
     studlab <- TE$studlab
@@ -546,8 +550,11 @@ discomb <- function(TE, seTE,
     #
     pairdata <- TE
     data <- TE
-    ##
-    TE <- TE$TE
+    #
+    if (is.null(attr(TE, "varnames")))
+      TE <- TE$TE
+    else
+      TE <- TE[[attr(TE, "varnames")[1]]]
   }
   else {
     is.pairwise <- FALSE